Rename /edit to /form
